moved router to views index
